# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PulseFlow API - AI WhatsApp front desk for beauty & wellness businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /business:
    get:
      operationId: getBusiness
      tags: [business]
      summary: Get business profile
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "404":
          description: Not found
    post:
      operationId: createBusiness
      tags: [business]
      summary: Create business profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [business]
      summary: Update business profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"

  # ─── Services ──────────────────────────────────────────────────────────────
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServicePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Customers ─────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer with visit history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ─── Bookings ──────────────────────────────────────────────────────────────
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List scheduled reminder jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderJob"
  /jobs/{id}/trigger:
    post:
      operationId: triggerJob
      tags: [jobs]
      summary: Manually trigger a reminder job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderJob"

  # ─── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Business:
      type: object
      required: [id, name, ownerName, phone, category, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        ownerName: { type: string }
        phone: { type: string }
        whatsappNumber: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        googleMapsLink: { type: ["string", "null"] }
        category: { type: string }
        description: { type: ["string", "null"] }
        timezone: { type: string }
        workingHours: {}
        cancellationPolicy: { type: ["string", "null"] }
        tokenPolicy: { type: ["string", "null"] }
        preferredTone: { type: string }
        reviewLink: { type: ["string", "null"] }
        whatsappVerifyToken: { type: ["string", "null"] }
        isOnboarded: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    BusinessInput:
      type: object
      required: [name, ownerName, phone, category]
      properties:
        name: { type: string }
        ownerName: { type: string }
        phone: { type: string }
        whatsappNumber: { type: string }
        city: { type: string }
        address: { type: string }
        googleMapsLink: { type: string }
        category: { type: string }
        description: { type: string }
        timezone: { type: string }
        workingHours: {}
        cancellationPolicy: { type: string }
        tokenPolicy: { type: string }
        preferredTone: { type: string }
        reviewLink: { type: string }
    BusinessPatch:
      type: object
      properties:
        name: { type: string }
        ownerName: { type: string }
        phone: { type: string }
        whatsappNumber: { type: string }
        city: { type: string }
        address: { type: string }
        googleMapsLink: { type: string }
        category: { type: string }
        description: { type: string }
        timezone: { type: string }
        workingHours: {}
        cancellationPolicy: { type: string }
        tokenPolicy: { type: string }
        preferredTone: { type: string }
        reviewLink: { type: string }
        whatsappVerifyToken: { type: string }
        isOnboarded: { type: boolean }

    # ─── Service ─────────────────────────────────────────────────────────────
    Service:
      type: object
      required: [id, businessId, name, price, durationMinutes, isActive, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        price: { type: number }
        durationMinutes: { type: integer }
        description: { type: ["string", "null"] }
        requiresConsultation: { type: boolean }
        requiresTokenAdvance: { type: boolean }
        repeatReminderDays: { type: ["integer", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    ServiceInput:
      type: object
      required: [name, price, durationMinutes]
      properties:
        name: { type: string }
        category: { type: string }
        price: { type: number }
        durationMinutes: { type: integer }
        description: { type: string }
        requiresConsultation: { type: boolean }
        requiresTokenAdvance: { type: boolean }
        repeatReminderDays: { type: integer }
        isActive: { type: boolean }
    ServicePatch:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        price: { type: number }
        durationMinutes: { type: integer }
        description: { type: string }
        requiresConsultation: { type: boolean }
        requiresTokenAdvance: { type: boolean }
        repeatReminderDays: { type: integer }
        isActive: { type: boolean }

    # ─── Customer ────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, businessId, name, phone, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        name: { type: string }
        phone: { type: string }
        notes: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        lastVisitAt: { type: ["string", "null"] }
        totalVisits: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    CustomerDetail:
      type: object
      required: [id, businessId, name, phone, createdAt, bookings]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        name: { type: string }
        phone: { type: string }
        notes: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        lastVisitAt: { type: ["string", "null"] }
        totalVisits: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        upcomingBooking:
          $ref: "#/components/schemas/Booking"
    CustomerList:
      type: object
      required: [customers, total]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total: { type: integer }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string }
        phone: { type: string }
        notes: { type: string }
        source: { type: string }
    CustomerPatch:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        notes: { type: string }
        source: { type: string }

    # ─── Booking ─────────────────────────────────────────────────────────────
    Booking:
      type: object
      required: [id, businessId, serviceId, bookingDate, startTime, endTime, status, source, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        customerId: { type: ["integer", "null"] }
        serviceId: { type: integer }
        bookingDate: { type: string, description: "YYYY-MM-DD" }
        startTime: { type: string, description: "HH:MM" }
        endTime: { type: string, description: "HH:MM" }
        status: { type: string, enum: [pending, confirmed, completed, cancelled, no_show, rescheduled] }
        source: { type: string, enum: [whatsapp, manual, dashboard] }
        notes: { type: ["string", "null"] }
        createdByAI: { type: boolean }
        reminder24hSent: { type: boolean }
        reminder2hSent: { type: boolean }
        reviewRequestSent: { type: boolean }
        repeatReminderScheduledAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        service:
          $ref: "#/components/schemas/Service"
...
    ReminderJob:
      type: object
      required: [id, businessId, type, scheduledFor, status, createdAt]
      properties:
        id: { type: integer }
        businessId: { type: integer }
        customerId: { type: ["integer", "null"] }
        bookingId: { type: ["integer", "null"] }
        type: { type: string, enum: [booking_confirmation, reminder_24h, reminder_2h, review_request, repeat_reminder] }
        scheduledFor: { type: string }
        status: { type: string, enum: [pending, sent, failed, cancelled] }
        payload: {}
        sentAt: { type: ["string", "null"] }
        createdAt: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        booking:
          $ref: "#/components/schemas/Booking"

    # ─── Dashboard ───────────────────────────────────────────────────────────
...