# lib/api-zod/src/generated/types/customerList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PulseFlow API - AI WhatsApp front desk for beauty & wellness businesses
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from './customer';

export interface CustomerList {
  customers: Customer[];
  total: number;
}