# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PulseFlow API - AI WhatsApp front desk for beauty & wellness businesses
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  businessId: number;
  name: string;
  /** @nullable */
  category?: string | null;
  price: number;
  durationMinutes: number;
  /** @nullable */
  description?: string | null;
  requiresConsultation?: boolean;
  requiresTokenAdvance?: boolean;
  /** @nullable */
  repeatReminderDays?: number | null;
  isActive: boolean;
  createdAt: string;
  updatedAt?: string;
}